requirements.txt
paho-mqtt
joblib
scikit-learn
streamlit
pandas
numpy